requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
httpx>=0.25.0
websockets>=12.0
supabase>=2.0.0
python-dateutil>=2.8.2 
supabase
python-multipart>=0.0.6
web3>=6.11.3
python-dotenv>=1.0.0
pandas>=2.1.3
